Fix big int... for reservation

+++ database/migrations/2019_01_15_104318_restaurant_tables.php
@@ -16,7 +16,6 @@ class RestaurantTables extends Migration
         Schema::create('reserveringen', function (Blueprint $table) {
 
             $table->bigInteger('reserveernummer'); //pk
-
             $table->integer('tijd');
             $table->date('datum', 20);
             $table->string('aantalGasten', 50);
@@ -71,7 +70,7 @@ class RestaurantTables extends Migration
         });
 
         Schema::create('tafelreserveringen', function (Blueprint $table) {
-            $table->integer('reserveernummer'); //fk, pk
+            $table->bigInteger('reserveernummer'); //fk, pk
             $table->integer('tafelnummer');     //fk, pk
             $table->dateTime('tijdin');
             $table->dateTime('tijduit');
